Switch default layout to component, not plugin
