+++ packages/rpv-default-layout/src/defaultLayout.less
@@ -36,4 +36,5 @@
 
 .rpv-default-layout-body {
     flex: 1;
+    overflow: auto;
 }
